requirements.txt
numpy
matplotlib
tqdm
pytorch
torchvision
tensorboard
gym
pybullet
pillow
moviepy
setproctitle
